backend/foodgram/requirements.txt
asgiref==3.5.2
Django==2.2.16
django-jsonfield==1.4.1
django-jsonfield-backport==1.0.5
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
fpdf==1.7.2
Pillow==9.2.0
pytz==2022.1
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
